# Unicode correctness