--- a/CONTRIBUTING.md
+++ b/CONTRIBUTING.md
@@ -3,3 +3,16 @@
 Thanks for your interest in contributing to `parsing`! We welcome new contributions regardless of your level of experience or familiarity with PureScript.
 
 Every library in the Contributors organization shares a simple handbook that helps new contributors get started. With that in mind, please [read the short contributing guide on purescript-contrib/governance](https://github.com/purescript-contrib/governance/blob/main/contributing.md) before contributing to this library.
+
+# Development
+
+This package includes a `spago-dev.dhall` which provides the dependencies
+for development and testing.
+
+## Testing
+
+To run the test suite:
+
+```
+spago -x spago-dev.dhall test
+```