@jamesdbrock 2022-01-11
Some more historical context: the `StringLike` typeclass was introduced in https://github.com/purescript-contrib/purescript-parsing/pull/36